fix: Health Connect activity-alias, VideoRenderer LiveKit, scores API alignment, bilan parsing

+++ app/src/main/java/health/telomer/android/feature/healthos/HealthOSScreen.kt
@@ -261,7 +261,7 @@ private fun InflammationBar(inflammation: InflammationResponse?) {
                         )
                 )
                 // Indicator
-                val fraction = (index.coerceIn(0, 100)) / 100f
+                val fraction = (index.coerceIn(0.0, 100.0) / 100.0).toFloat()
                 Box(
                     modifier = Modifier
                         .fillMaxWidth(fraction)
